Move most of code for getting latest tag/version and calling publish_release to postpublish_utils script in top-level dir
