+++ package.json
@@ -13,6 +13,10 @@
       "mnemonic": "concert load couple harbor equip island argue ramp clarify fence smart topic"
     },
     "devDependencies": {
-        "lerna": "^2.5.1"
+        "lerna": "^2.5.1",
+        "async-child-process": "^1.1.1",
+        "semver-sort": "^0.0.4",
+        "publish-release": "0xproject/publish-release",
+        "es6-promisify": "^5.0.0"
     }
 }
