+++ packages/0x.js/package.json
@@ -70,7 +70,6 @@
     "prompt": "^1.0.0",
     "request": "^2.81.0",
     "request-promise-native": "^1.0.4",
-    "semver-sort": "^0.0.4",
     "shx": "^0.2.2",
     "sinon": "^4.0.0",
     "source-map-support": "^0.5.0",
@@ -90,14 +89,12 @@
     "bignumber.js": "~4.1.0",
     "bn.js": "4.11.8",
     "compare-versions": "^3.0.1",
-    "es6-promisify": "^5.0.0",
     "ethereumjs-abi": "^0.6.4",
     "ethereumjs-blockstream": "^2.0.6",
     "ethereumjs-util": "^5.1.1",
     "find-versions": "^2.0.0",
     "js-sha3": "^0.6.1",
     "lodash": "^4.17.4",
-    "publish-release": "0xproject/publish-release",
     "uuid": "^3.1.0",
     "web3": "^0.20.0"
   }
